requirements.txt
asgiref==3.7.2
Django==3.0.14
django-ckeditor==6.1.0
django-debug-toolbar==2.2
django-js-asset==2.0.0
django-mptt==0.11.0
Pillow==8.1.2
python-dotenv==0.17.0
pytz==2024.2
sqlparse==0.4.4
typing-extensions==4.7.1
